move parallelize function to a separate file
