+++ examples/fabric/tensor_parallel/train.py
@@ -1,11 +1,13 @@
 import lightning as L
 import torch
 import torch.nn.functional as F
-from data import RandomTokenDataset
-from lightning.fabric.strategies import ModelParallelStrategy
-from model import ModelArgs, Transformer, parallelize
 from torch.distributed.tensor.parallel import loss_parallel
 from torch.utils.data import DataLoader
+from lightning.fabric.strategies import ModelParallelStrategy
+
+from data import RandomTokenDataset
+from model import ModelArgs, Transformer
+from parallelism import parallelize
 
 
 def train():
